+++ portal/lib/app-version.ts
@@ -1,2 +1,2 @@
-export const APP_VERSION = "3.8.34";
-export const APP_BUILD_TIME = "2026-04-30";
+export const APP_VERSION = "3.8.36";
+export const APP_BUILD_TIME = "2026-05-03";